numerique-main/requirements.txt
numpy
pandas
matplotlib
scipy # for reading/writing wav files
pillow
seaborn
scikit_learn
openpyxl
ipythontutor

# tooling
jupyter
jupytext
jupyterlab-courselevels
jupyterlab-myst
ipywidgets
ipympl

# for authors only
# nbprune
# jupyter-book
# sphinx-exercise

# for updating nbhosting toc from jupyter book
ruamel.yaml
